+++ sisl/physics/tests/test_hamiltonian.py
@@ -812,18 +812,48 @@ class TestHamiltonian:
         bz = BandStructure.param_circle(H, 20, 0.01, [0, 0, 1], [1/3] * 3)
         berry_phase(bz)
 
-    def test_berry_phase_zak_x(self):
+    def test_berry_phase_zak_x_topological(self):
         # SSH model, topological cell
-        g = Geometry([[-.6, 0, 0], [0.6, 0, 0]], Atom(1, 1.001), sc=[2, 10, 10])
+        # |t2| < |t1|
+        g = Geometry([[0, 0, 0], [1.2, 0, 0]], Atom(1, 1.001), sc=[2, 10, 10])
         g.set_nsc([3, 1, 1])
         H = Hamiltonian(g)
         H.construct([(0.1, 1.0, 1.5), (0, 1., 0.5)])
         # Contour
-        k = np.linspace(0.0, 1.0, 101)
-        K = np.zeros([k.size, 3])
-        K[:, 0] = k
-        bz = BrillouinZone(H, K)
+        def func(parent, N, i):
+            return [i/N, 0, 0]
+        bz = BrillouinZone.parametrize(H, func, 101)
         assert np.allclose(np.abs(berry_phase(bz, sub=0, method='zak')), np.pi)
+        # Just to do the other branch
+        berry_phase(bz, method='zak')
+
+    def test_berry_phase_zak_x_topological_non_orthogonal(self):
+        # SSH model, topological cell
+        # |t2| < |t1|
+        g = Geometry([[0, 0, 0], [1.2, 0, 0]], Atom(1, 1.001), sc=[2, 10, 10])
+        g.set_nsc([3, 1, 1])
+        H = Hamiltonian(g, orthogonal=False)
+        H.construct([(0.1, 1.0, 1.5), ((0, 1), (1., 0.25), (0.5, 0.1))])
+        # Contour
+        def func(parent, N, i):
+            return [i/N, 0, 0]
+        bz = BrillouinZone.parametrize(H, func, 101)
+        assert np.allclose(np.abs(berry_phase(bz, sub=0, method='zak')), np.pi)
+        # Just to do the other branch
+        berry_phase(bz, method='zak')
+
+    def test_berry_phase_zak_x_trivial(self):
+        # SSH model, trivial cell
+        # |t2| > |t1|
+        g = Geometry([[0, 0, 0], [1.2, 0, 0]], Atom(1, 1.001), sc=[2, 10, 10])
+        g.set_nsc([3, 1, 1])
+        H = Hamiltonian(g)
+        H.construct([(0.1, 1.0, 1.5), (0, 0.5, 1.)])
+        # Contour
+        def func(parent, N, i):
+            return [i/N, 0, 0]
+        bz = BrillouinZone.parametrize(H, func, 101)
+        assert np.allclose(np.abs(berry_phase(bz, sub=0, method='zak')), 0.)
         # Just to do the other branch
         berry_phase(bz, method='zak')
 
@@ -849,7 +879,7 @@ class TestHamiltonian:
         H.construct([(0.1, 1.0, 1.5), (0, 1., 0.5)])
         # Contour
         def func(parent, N, i):
-            return [0, i/N, 0]
+            return [i/N, 0, 0]
         bz = BrillouinZone.parametrize(H, func, 101)
         zak = berry_phase(bz, sub=0, method='zak')
         assert np.allclose(np.abs(zak), np.pi)